Moved library and main into separate files
